구매 비누 외형 건수 모두 0일 때 극소량 값 반환

+++ src/main/java/com/bumsoap/store/dto/ShapeLabelCount.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @Setter
 public class ShapeLabelCount {
     private String shapeLabel;
-    private Long count;
+    private float count;
 }
